+++ src/main/java/com/hyundaiautoever/HEAT/v1/service/UserService.java
@@ -123,7 +123,7 @@ public class UserService {
         //유저 이름 업데이트
         user.setUserName(updateUserDto.getUserName());
         //유저 프로필 사진 업데이트
-        if (userProfileImage.isPresent()) {
+        if (!userProfileImage.get().isEmpty()) {
             //기존 이미지 삭제
             s3Service.removeS3File(user.getProfileImageUrl());
             String userProfileImageUrl = s3Service.uploadUserProfileImage(userProfileImage.get());
@@ -158,7 +158,9 @@ public class UserService {
      */
     public void deleteUser(Long userAccountNo) {
         User user = userRepository.findByUserAccountNo(userAccountNo);
-        s3Service.removeS3File(user.getProfileImageUrl());
+        if (user.getProfileImageUrl() != null) {
+            s3Service.removeS3File(user.getProfileImageUrl());
+        }
         userRepository.deleteByUserAccountNo(userAccountNo);
     }
 }
